moved entry point of the algorithm
